// src/interfaces/BotonPanicoInterface.ts
export interface BotonPanicoInterface {
  altitude: number;
  asset: string;
  battery: number;
  code: string;
  course: number;
  customerid: string;
  customername: string;
  date: string;
  direction: string;
  humidity: number;
  ignition: boolean;
  latitude: number;
  longitude: number;
  odometer: number;
  serialNumber: string;
  shipment: string;
  speed: number;
  temperature: number;
  vehicleType?: string;
  vehicleBrand?: string;
  vehicleModel?: string;

}